Move StringArgument to a separate file
